+++ NEWS.md
@@ -1,5 +1,14 @@
 # fsr 1.0.2.9000 (development version)
 
+## Major changes
+
+- Defined a new hierarchical structure for defining the spatial plateau data types:
+  - Added two new spatial plateau data types: `pcollection` and `pcomposition`.
+  - Redefined the class named `pgeometry`.
+  - Defined three spatial plateau data types that inherit from `pgeometry`: `ppoint`, `pline`, and `pregion`.
+  - Specified validity methods for all classes according to the rules of the Spatial Plateau Algebra
+  - These modifications are based on the pull requests performed by Pedro V. A. B. de Venâncio, who will be added as a contributor of the package `fsr` ([#26 - pedbrgs](https://github.com/accarniel/fsr/pull/26) and [#27 - pedbrgs](https://github.com/accarniel/fsr/pull/27))
+
 ## Minor changes
 
 - Better organization of the source code by creating more R scripts and rearranging the functions and classes in them.